Cleaned and refactored app.py by splitting routes
