+++ Templates/asset_form.html
@@ -4,10 +4,10 @@
   href="{{ url_for('static', filename='css/form.css') }}"
 />
 {% set title = "Add Asset" if not asset else "Edit Asset" %} {% set
-form_action_url = url_for('add_asset') if not asset else url_for('edit_asset',
+form_action_url = url_for('assets.add_asset') if not asset else url_for('assets.edit_asset',
 asset_id=asset.id) %} {% if request_form %} {% set form_action_url =
-url_for('add_asset') %} {% endif %} {% if asset %} {% set title = "Edit Asset"
-%} {% set form_action_url = url_for('edit_asset', asset_id=asset.id) %} {% endif
+url_for('assets.add_asset') %} {% endif %} {% if asset %} {% set title = "Edit Asset"
+%} {% set form_action_url = url_for('assets.edit_asset', asset_id=asset.id) %} {% endif
 %}
 
 <body>
@@ -143,7 +143,7 @@ url_for('add_asset') %} {% endif %} {% if asset %} {% set title = "Edit Asset"
           >
         </div>
         <div class="form-actions">
-          <a href="{{ url_for('list_assets') }}" class="btn btn-secondary"
+          <a href="{{ url_for('assets.list_assets') }}" class="btn btn-secondary"
             >Cancel</a
           >
           <button type="submit" class="btn btn-primary">
