+++ Templates/assets_list.html
@@ -24,7 +24,7 @@
       </div>
       <div class="table-header">
         <h2 class="title">All Assets</h2>
-        <a href="{{ url_for('add_asset') }}" class="btn-primary">
+        <a href="{{ url_for('assets.add_asset') }}" class="btn-primary">
           <svg
             class="icon"
             fill="none"
@@ -75,7 +75,7 @@
               <td class = "table-data">{{ asset.creator.name if asset.creator else 'N/A' }}</td>
               <td class="table-data">
                 <a
-                  href="{{ url_for('edit_asset', asset_id=asset.id) }}"
+                  href="{{ url_for('assets.edit_asset', asset_id=asset.id) }}"
                   class="edit"
                   title="Edit"
                 >
@@ -83,7 +83,7 @@
                 </a>
                 {% if current_user.role == 'admin' %}
                 <form
-                  action="{{ url_for('delete_asset', asset_id=asset.id) }}"
+                  action="{{ url_for('assets.delete_asset', asset_id=asset.id) }}"
                   method="POST"
                   class="inline-block"
                   onsubmit="return confirm('Are you sure you want to delete this asset? This action cannot be undone.');"
