+++ Templates/categories_list.html
@@ -18,7 +18,7 @@
     <div class="content-header">
       <h2 class="title">Manage Categories</h2>
       {% if current_user.role == 'admin' %}
-      <a href="{{ url_for('add_category') }}" class="btn-primary">
+      <a href="{{ url_for('assets.add_category') }}" class="btn-primary">
         <svg
           class="icon"
           fill="none"
@@ -58,7 +58,7 @@
         <div class="category-actions">
           {% if current_user.role == 'admin' %}
           <a
-            href="{{ url_for('edit_category', category_id=category.id) }}"
+            href="{{ url_for('assets.edit_category', category_id=category.id) }}"
             class="edit"
           >
             <svg
@@ -83,7 +83,7 @@
             <span class="icon-only">Edit</span>
           </a>
           <form
-            action="{{ url_for('delete_category', category_id=category.id) }}"
+            action="{{ url_for('assets.delete_category', category_id=category.id) }}"
             method="POST"
             onsubmit="return confirm('Are you sure you want to delete this category? This may affect existing assets if not handled properly.');"
           >
