+++ Templates/profile_form.html
@@ -22,7 +22,7 @@
 
       <form
         method="POST"
-        action="{{ url_for('edit_profile') }}"
+        action="{{ url_for('auth.edit_profile') }}"
       >
         <div class="form-grid">
           <div class="form-group">
@@ -116,7 +116,7 @@
           </div>
         </div>
         <div class="form-actions">
-          <a href="{{ url_for('dashboard') }}" class="btn btn-secondary"
+          <a href="{{ url_for('assets.dashboard') }}" class="btn btn-secondary"
             >Cancel</a
           >
           <button type="submit" class="btn btn-primary">Save Changes</button>
