+++ Templates/user_form.html
@@ -24,8 +24,8 @@
       </div>
       <h2 class="section-title {% if form_mode == 'edit' %} edit {% else %} add {% endif %}">User Information</h2>
 
-      {% set form_action_url = url_for('add_user') if form_mode == 'add' else
-      url_for('edit_user', user_id=user_to_edit.id) %} {% if form_mode == 'edit'
+      {% set form_action_url = url_for('admin.add_user') if form_mode == 'add' else
+      url_for('admin.edit_user', user_id=user_to_edit.id) %} {% if form_mode == 'edit'
       and user_to_edit %}
       <p class="info">
         User password changes are handled by the user via their profile page or
@@ -110,7 +110,7 @@
           {% endif %}
         </div>
         <div class="form-actions">
-          <a href="{{ url_for('list_users') }}" class="btn btn-secondary"
+          <a href="{{ url_for('admin.list_users') }}" class="btn btn-secondary"
             >Cancel</a
           >
           <button type="submit" class="btn btn-primary">
