+++ Templates/user_list.html
@@ -24,7 +24,7 @@
       </div>
       <div class="table-header">
         <h2 class="title">All Users</h2>
-        <a href="{{ url_for('add_user') }}" class="btn-primary">
+        <a href="{{ url_for('admin.add_user') }}" class="btn-primary">
           <svg
             class="icon"
             fill="none"
@@ -73,7 +73,7 @@
               <td class="table-data">{{ user_item.registration_date.strftime('%Y-%m-%d') if user_item.registration_date else 'N/A' }}</td>
               <td class="table-data">
                 <a
-                  href="{{ url_for('edit_user', user_id=user_item.id) }}"
+                  href="{{ url_for('admin.edit_user', user_id=user_item.id) }}"
                   class="edit"
                   title="Edit"
                 >
@@ -81,7 +81,7 @@
                 </a>
                 {% if current_user.id != user_item.id %}
                 <form
-                  action="{{ url_for('delete_user', user_id=user_item.id) }}"
+                  action="{{ url_for('admin.delete_user', user_id=user_item.id) }}"
                   method="POST"
                   class="inline-block"
                   onsubmit="return confirm('Are you sure you want to delete user {{ user_item.username }}? This action cannot be undone.');"
